split fixaas stuff from q3map2
