+++ tools/quake3/q3map2/main.c
@@ -91,86 +91,6 @@ static void ExitQ3Map( void ){
 	}
 }
 
-static int MD4BlockChecksum( void * buffer, int length ) {
-	unsigned char digest[16];
-	int checksum;
-
-	md4_get_digest( buffer, length, digest );
-	/* I suppose it has to be done that way for legacy reasons? */
-	checksum = digest[0] & ( digest[1] << 8 ) & ( digest[2] << 16 ) & ( digest[3] << 24 );
-	checksum ^= digest[4] & ( digest[5] << 8 ) & ( digest[6] << 16 ) & ( digest[7] << 24 );
-	checksum ^= digest[8] & ( digest[9] << 8 ) & ( digest[10] << 16 ) & ( digest[11] << 24 );
-	checksum ^= digest[12] & ( digest[13] << 8 ) & ( digest[14] << 16 ) & ( digest[15] << 24 );
-	return checksum;
-}
-
-/*
-   FixAAS()
-   resets an aas checksum to match the given BSP
- */
-
-int FixAAS( int argc, char **argv ){
-	int length, checksum;
-	void        *buffer;
-	FILE        *file;
-	char aas[ 1024 ], **ext;
-	char        *exts[] =
-	{
-		".aas",
-		"_b0.aas",
-		"_b1.aas",
-		NULL
-	};
-
-
-	/* arg checking */
-	if ( argc < 2 ) {
-		Sys_Printf( "Usage: q3map -fixaas [-v] <mapname>\n" );
-		return 0;
-	}
-
-	/* do some path mangling */
-	strcpy( source, ExpandArg( argv[ argc - 1 ] ) );
-	StripExtension( source );
-	DefaultExtension( source, ".bsp" );
-
-	/* note it */
-	Sys_Printf( "--- FixAAS ---\n" );
-
-	/* load the bsp */
-	Sys_Printf( "Loading %s\n", source );
-	length = LoadFile( source, &buffer );
-
-	/* create bsp checksum */
-	Sys_Printf( "Creating checksum...\n" );
-	checksum = LittleLong( MD4BlockChecksum( buffer, length ) );
-
-	/* write checksum to aas */
-	ext = exts;
-	while ( *ext )
-	{
-		/* mangle name */
-		strcpy( aas, source );
-		StripExtension( aas );
-		strcat( aas, *ext );
-		Sys_Printf( "Trying %s\n", aas );
-		ext++;
-
-		/* fix it */
-		file = fopen( aas, "r+b" );
-		if ( !file ) {
-			continue;
-		}
-		if ( fwrite( &checksum, 4, 1, file ) != 1 ) {
-			Error( "Error writing checksum to %s", aas );
-		}
-		fclose( file );
-	}
-
-	/* return to sender */
-	return 0;
-}
-
 
 /*
    main()
@@ -273,7 +193,7 @@ int main( int argc, char **argv ){
 
 	/* fixaas */
 	if ( !strcmp( argv[ 1 ], "-fixaas" ) ) {
-		r = FixAAS( argc - 1, argv + 1 );
+		r = FixAASMain( argc - 1, argv + 1 );
 	}
 
 	/* analyze */
