+++ tools/quake3/q3map2/q3map2.vcproj
@@ -243,6 +243,10 @@
 				>
 			</File>
 			<File
+				RelativePath=".\fixaas.c"
+				>
+			</File>
+			<File
 				RelativePath=".\fog.c"
 				>
 			</File>
