+++ tools/quake3/q3map2/q3map2.vcxproj
@@ -179,6 +179,7 @@
     <ClCompile Include="bsp_info.c" />
     <ClCompile Include="decals.c" />
     <ClCompile Include="facebsp.c" />
+    <ClCompile Include="fixaas.c" />
     <ClCompile Include="fog.c" />
     <ClCompile Include="leakfile.c" />
     <ClCompile Include="map.c" />
