+++ tools/quake3/q3map2/q3map2.vcxproj.filters
@@ -65,6 +65,9 @@
     <ClCompile Include="facebsp.c">
       <Filter>src</Filter>
     </ClCompile>
+    <ClCompile Include="fixaas.c">
+      <Filter>src</Filter>
+    </ClCompile>
     <ClCompile Include="fog.c">
       <Filter>src</Filter>
     </ClCompile>
